Continue converting ooDialog to the C++ API
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@5386 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ extensions/platform/windows/oodialog/ooDialog.hpp
@@ -386,8 +386,8 @@ typedef struct {
 
 typedef struct
 {
-   void * previous;
-   INT TableEntry;
+   void              *previous;
+   size_t             TableEntry;
    MESSAGETABLEENTRY *MsgTab;
    DATATABLEENTRY    *DataTab;
    BITMAPTABLEENTRY  *BmpTab;
